POPR-179: added types workspace
