+++ backend/package.json
@@ -30,7 +30,8 @@
     "mongoose": "^6.5.2",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "^7.2.0"
+    "rxjs": "^7.2.0",
+    "@meethead/types": "^1.0.0"
   },
   "devDependencies": {
     "@nestjs/cli": "^9.0.0",
