--- backend/src/contacts/dto/create-contact.dto.ts
@@ -1,8 +1,0 @@
-export class CreateContactDto {
-  name: string;
-  profilePictureUrl: string;
-  companyName: string;
-  companyLogoUrl: string;
-  companyPosition: string;
-  linkedInURL: string;
-}
