--- backend/src/contacts/dto/response-contact.dto.ts
@@ -1,8 +1,0 @@
-export class ResponseContactDTO {
-  name: string;
-  profilePictureS3Id: string;
-  companyName: string;
-  logoS3Id: string;
-  companyPosition: string;
-  linkedInURL: string;
-}
