+++ frontend/app/package.json
@@ -59,6 +59,7 @@
   "dependencies": {
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
+    "@meethead/types": "^1.0.0",
     "@meethead/ui": "^1.0.0",
     "@mui/material": "^5.10.3",
     "@mui/styled-engine-sc": "^5.10.3",
