+++ frontend/extension/package.json
@@ -57,6 +57,8 @@
     ]
   },
   "dependencies": {
+    "@meethead/types": "^1.0.0",
+    "@meethead/ui": "^1.0.0",
     "axios": "^0.27.2",
     "classnames": "^2.3.1",
     "i18next": "^21.8.9",
@@ -69,8 +71,7 @@
     "react-query": "^3.39.1",
     "react-router-dom": "^6.3.0",
     "styled-components": "^5.3.5",
-    "web-vitals": "^2.1.4",
-    "@meethead/ui": "^1.0.0"
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
     "@testing-library/react": "^13.3.0",
