+++ frontend/ui/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
+    "@meethead/types": "^1.0.0",
     "@mui/material": "^5.10.3",
     "@mui/styled-engine-sc": "^5.10.3",
     "@rollup/plugin-alias": "^3.1.9",
