+++ package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "ISC",
   "workspaces": [
+    "types",
     "backend",
     "frontend/*"
   ],
