+++ types/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@meethead/types",
+  "version": "1.0.0",
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "rm -rf dist && rollup --config",
+    "start": "rollup --config -w"
+  },
+  "dependencies": {
+    "@rollup/plugin-alias": "^3.1.9",
+    "rollup": "^2.60.2",
+    "rollup-plugin-typescript2": "^0.31.1",
+    "tslib": "^2.3.1",
+    "typescript": "^4.5.2"
+  }
+}
