+++ types/rollup.config.js
@@ -1,0 +1,14 @@
+import typescript from "rollup-plugin-typescript2";
+
+export default {
+  input: ["src/index.ts"],
+  output: [
+    {
+      dir: "dist",
+      entryFileNames: "[name].js",
+      format: "cjs",
+      exports: "named",
+    },
+  ],
+  plugins: [typescript()],
+};
