+++ types/src/User/index.ts
@@ -1,4 +1,4 @@
-export class CreateUserDto {
+export class CreateUserDTO {
   public email: string;
   public password: string;
   public contacts: [];
