+++ types/src/index.ts
@@ -1,0 +1,2 @@
+export { CreateContactDTO, ResponseContactDTO } from "./Contact";
+export { CreateUserDTO } from "./User";
